python-agent/pyproject.toml
[project]
name = "bom-agent-service"
version = "0.1.0"
description = "CrewAI-powered BOM sourcing agent with OpenAI-compatible API"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "crewai>=0.86.0",
    "openai>=1.0.0",
    "pydantic>=2.0.0",
    "httpx>=0.27.0",
    "rich>=13.0.0",
    "click>=8.0.0",
    "pyyaml>=6.0",
]

[project.scripts]
sourcing = "bom_agent_service.cli:main"
sourcing-server = "bom_agent_service.server:main"
sourcing-demo = "bom_agent_service.demo:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/bom_agent_service"]

[dependency-groups]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "asgi-lifespan>=2.1.0",
    "python-dotenv>=1.0.0",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"
filterwarnings = [
    "ignore::DeprecationWarning",
    "ignore::pytest.PytestRemovedIn9Warning",
]
